Split upgrade tests as we hit the timeout for our CI pipeline
